do not use long

+++ src/main.cpp
@@ -27,7 +27,7 @@ int main(int argc, char *argv[]) {
     clock_t start = clock();
 
     // KNN algorithm
-    long* h_sorted_indices = knn(h_documents, h_queries, D, N, Q);
+    int* h_sorted_indices = knn(h_documents, h_queries, D, N, Q);
 
     // Measure elapsed time
     clock_t end = clock();
