+++ src/test_knn_cuda.cu
@@ -23,7 +23,7 @@ int main(int argc, char *argv[]) {
     float *h_queries = (float *)malloc(Q * D * sizeof(float));
     float *h_distances = (float *)malloc(Q * N * D * sizeof(float));
     float *h_agg_distances = (float *)malloc(Q * N * sizeof(float));
-    int *h_indices = (int *)malloc(Q * N * sizeof(long)); // Indices array to store the output of kSelectKernel
+    int *h_indices = (int *)malloc(Q * N * sizeof(int )); // Indices array to store the output of kSelectKernel
 
     // Initialize data with random values
     srand(time(NULL));
@@ -40,7 +40,7 @@ int main(int argc, char *argv[]) {
     cudaMalloc(&d_queries, Q * D * sizeof(float));
     cudaMalloc(&d_distances, Q * N * D * sizeof(float));
     cudaMalloc(&d_agg_distances, Q * N * sizeof(float));
-    cudaMalloc(&d_indices, Q * N * sizeof(long)); // Device memory for indices
+    cudaMalloc(&d_indices, Q * N * sizeof(int )); // Device memory for indices
 
     // Copy data from host to device
     cudaMemcpy(d_documents, h_documents, N * D * sizeof(float), cudaMemcpyHostToDevice);
@@ -80,10 +80,10 @@ int main(int argc, char *argv[]) {
     kSelectKernel<<<kSelectBlocksPerGrid, kSelectThreadsPerBlock>>>(d_agg_distances, d_indices, Q, N, K);
     
     // Copy the sorted indices back to the host
-    cudaMemcpy(h_indices, d_indices, Q * N * sizeof(long), cudaMemcpyDeviceToHost);
+    cudaMemcpy(h_indices, d_indices, Q * N * sizeof(int ), cudaMemcpyDeviceToHost);
 
     // Copy the result back to host
-    long* h_sorted_indices = argsort(h_agg_distances, Q, N);
+    int * h_sorted_indices = argsort(h_agg_distances, Q, N);
 
     // Measure elapsed time
     clock_t end = clock();
@@ -109,7 +109,7 @@ int main(int argc, char *argv[]) {
     // Perform the same operations on the CPU
     computeL1Distance(h_documents, h_queries, h_distances_cpu, D, N, Q);
     sumOverLastDim(h_distances_cpu, h_agg_distances_cpu, D, N, Q);
-    long* h_sorted_indices_cpu = argsort(h_agg_distances_cpu, Q, N);
+    int * h_sorted_indices_cpu = argsort(h_agg_distances_cpu, Q, N);
     
     // Verify the distances by comparing the GPU and CPU results
     printf("\nVerifying distance computation...\n");
